Fix compiler warnings about IEquatable<T> by making classes sealed
